fix: publish spot and point shadow pages as coherent grids
- The page table only publishes a spot plane or a cube face once every page in it is resident, rendered, and clean, so a sampler never mixes fresh and stale light-space contents into page-shaped lighting seams
- Point faces drain as whole units under an expanded budget, receiver-demanded faces first; a sub-texel light move keeps the last coherent face visible while it refreshes
- A moved caster dirties the armed projective grids wholesale instead of projecting every cooked leaf box through six faces on the CPU; directional levels keep per-page precision
- Wind dirtying now checks the active world for wind instances, and point sampling takes a receiver depth bias to stop self-shadow acne

+++ engine/crates/rendering/src/renderer/mod.rs
@@ -659,8 +659,8 @@ pub struct Renderer {
     vsm_demanded: Vec<u32>,
     /// The spot matrix the resident spot pages were rendered under.
     vsm_spot_matrix: [f32; 16],
-    /// The point light's position + far of the pages on the atlas; a change invalidates every face.
-    vsm_point_key: [f32; 4],
+    /// The point light's exact position + far key for pages on the atlas.
+    vsm_point_key: [u32; 4],
     /// Per-world interaction field buffers (header + damped-oscillator texel cascades).
     interaction_fields: std::collections::HashMap<u64, crate::Buffer>,
     /// Impulses staged for this frame's interaction-field step.
